refactor: Convert mesh generation to Cube<u8> and add comprehensive tests
This commit completes the conversion from i32 to u8 for material IDs in the
mesh generation system and adds extensive test infrastructure to validate
mesh generation correctness.

## Changes

### Test Infrastructure
- Add crates/cube/tests/test_models.rs with CSM-based test models at depths 0-3
  - Single leaf cube, octa cubes, extended depth models
  - Special cases: all solid, all empty, checkerboard patterns
  - CSM parser integration with i32→u8 conversion layer

- Add crates/cube/tests/mesh_generation_tests.rs with 17 comprehensive tests
  - Vertex position and normal validation
  - Face count tests for various octree structures
  - Multi-depth mesh generation tests
  - Demonstrates current mesh generation issues (3 faces instead of 6)

### Mesh Generation System
- Convert ColorMapper trait to use u8 instead of i32
  - Update HsvColorMapper, PaletteColorMapper, VoxColorMapper
  - Fix literal overflow with u32 intermediate cast

- Update NeighborGrid to use Cube<u8> and [u8; 4] border materials

- Update generate_face_mesh signature to use Cube<u8> and u8 materials

- Invert face traversal logic in visit_faces.rs
  - Old: Look FROM empty voxels TOWARD solid (broken)
  - New: Look FROM solid voxels TOWARD empty (correct semantics)
  - Handle None neighbors (outside grid) as empty
  - Update Face::DIRECTIONS mappings to match inverted logic
  - Preserve old implementation as visit_faces_old (deprecated)

### Core Cube Type Support
- Add Cube<u8>::id() implementation alongside existing Cube<i32>::id()
- Fix method ambiguity in tests using fully qualified syntax
  - Pattern: <Cube<i32>>::id(&cube) for explicit type resolution

### World Integration
- Add runtime conversion from Cube<i32> (terrain) to Cube<u8> (mesh gen)
- Update MaterialColorMapper to handle u8→i32 cast for material lookup
- Move existing inline tests from generator.rs to dedicated test file

### Other Updates
- Update orthographic renderer to cast i32 to u8 for color mapper
- Update editor mesh_sync to use u8 conversion
- Add world crate tests for mesh generation validation

## Test Results

All code compiles successfully. Mesh generation tests execute and demonstrate
the current issues:
- Single leaf cube: Expected 6 faces, got 3 (missing top/right/back)
- Octa cube depth 1: Expected faces, got 0
- All solid cube: Expected 6 boundary faces, got 3

Tests successfully demonstrate that mesh generation is currently broken and
provide a solid foundation for debugging and fixing the traversal logic.

+++ crates/cube/src/render/orthographic.rs
@@ -315,7 +315,7 @@ fn extract_voxels_from_octree(
                 let size = voxel_size * scale_factor as f32;
 
                 // Get color and convert to u8
-                let rgb = mapper.map(*value);
+                let rgb = mapper.map(*value as u8);
                 let color = [
                     (rgb[0] * 255.0) as u8,
                     (rgb[1] * 255.0) as u8,
@@ -460,7 +460,7 @@ fn draw_voxel_2d(
     let (u_range, v_range, _depth) = project_voxel(position, size, direction);
 
     // Get color from mapper
-    let color = mapper.map(value);
+    let color = mapper.map(value as u8);
     let color_u8 = [
         (color[0] * 255.0) as u8,
         (color[1] * 255.0) as u8,
